Reorganizing project structure
