+++ JRA12L/Core/ChessTable.cs
@@ -1,4 +1,8 @@
-namespace JRA12L;
+using JRA12L.Model;
+using JRA12L.Model.Figures;
+using JRA12L.View;
+
+namespace JRA12L.Core;
 
 public class ChessTable : ITable
 {
