+++ JRA12L/View/ConsoleUserInput.cs
@@ -1,4 +1,4 @@
-namespace JRA12L;
+namespace JRA12L.View;
 
 public class ConsoleUserInput : IUserInput
 {
